删除了不合理的基础结构 Signed-off-by: leafspace <leafspace_zhang@163.com>

+++ 程序代码/重构中/Windows Hello - VPR/Windows Hello - VPR/MediaControl/Common/CommonStruct.h
@@ -2,10 +2,10 @@
 
 struct VoiceParagraph
 {
-	unsigned long begin;                                                    // 语音段落开始点
-	unsigned long end;                                                      // 语音段落结束点
-	unsigned long voiceLength;                                              // 语音段落长度
-	VoiceParagraph(unsigned long be, unsigned long en, unsigned long vo) {  // 构造初始化
+	unsigned int begin;                                                               // 语音段落开始点
+	unsigned int end;                                                                 // 语音段落结束点
+	unsigned int voiceLength;                                                         // 语音段落长度
+	VoiceParagraph(unsigned int be, unsigned int en, unsigned int vo) {               // 构造初始化
 		begin = be;
 		end = en;
 		voiceLength = vo;
